fix:컨벤션에 맞게 페이지에 제작
