--- src/pages/quote/create/page.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const page = () => {
-  return <div>page</div>;
-};
